Create  DNA Fragment Analysis with R
# DNA Fragment Analysis with R by Albert Mahaffey

This R script performs a series of statistical tests and visualizations on DNA fragment data. It takes in a CSV file of reference data and a CSV file of sample data, each consisting of DNA fragment information from a gel image analysis. 

The program estimates the fragment sizes based on a DNA ladder, applies Z-score normalization to the calibrated volumes, and then compares the sample data to the reference data using ANOVA, outputting a p-value that can be used to determine if the sample matches the reference.

The script also generates several types of plots for visualizing the data, including histograms, boxplots, scatter plots, heatmaps, correlation plots, and line plots.