Create DNA_Fragment_Analysis_Of_Variance
# Fragment Size Analysis with R

Author: Albert Mahaffey

This repository hosts an R script designed by Albert Mahaffey to perform fragment size analysis, which is often useful in genomics and related biological fields. The script applies a variety of statistical tests, including ANOVA and Cohen's d, and generates visualizations to aid in data interpretation.

## Overview

The primary objective of this script is to analyze and compare the fragment sizes of different sample sets. It's particularly designed to handle data related to genomic bands, such as those derived from Gel Electrophoresis experiments. 

Starting by estimating fragment sizes based on a reference ladder, the script then applies ANOVA tests to compare fragment size distributions between a reference sample and other samples. The script further calculates the Cohen's d effect size and performs a power analysis to ensure the sample size's adequacy.

To facilitate data interpretation, visualizations are created for each sample set and comparison, including boxplots for comparing fragment sizes, histograms for observing fragment sizes' distribution within a sample, and Q-Q plots for checking normality.

## Dependencies

This script utilizes various R libraries:

- readr: for reading csv files
- dplyr and tidyr: for data manipulation
- pheatmap: for generating heatmaps
- ggplot2: for creating visualizations
- corrplot: for plotting correlation matrices
- viridis: for additional color palettes
- effsize: for effect size calculation
- pwr: for power analysis
- BayesFactor: for Bayesian analysis

## Input Data

The input data should be in CSV format, including fields such as:

- Lane: the lane of the gel electrophoresis
- Band: the band of the gel electrophoresis
- Rf: the relative front (Rf) value representing a band's migration distance
- Sample: the name of the sample

A key assumption in this script is the existence of a reference ladder (a lane containing DNA fragments of known size) to estimate other fragments' sizes.

## Output

 Script generates a series of visualizations for each comparison scenario, including boxplots, histograms, and Q-Q plots. It also provides the p-value, effect size (Cohen's d), and power for each comparison scenario.

Please adjust the file paths and other parameters in the script to match your data and analysis requirements.